# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Index Card API",
    "description" : "This is a simple server for loading and saving index cards",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://dev.index-cards.com/v1",
    "description" : "Dev Server"
  }, {
    "url" : "https://prod.index-cards.com/v1",
    "description" : "Prod Server"
  } ],
...
  "paths" : {
    "/users" : {
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Create a new user",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/users_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Successfully created new user"
          },
          "409" : {
            "description" : "Username already exists"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/users/{userId}" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Get information about the user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Sucessful pull of user info",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "404" : {
            "description" : "User not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "users" ],
        "summary" : "Delete user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "successfully deleted user"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "404" : {
            "description" : "user not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "users" ],
        "summary" : "update selected user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "requestBody" : {
          "description" : "New information to update user",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/users_userId_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "sucessfully updated user"
          },
          "401" : {
            "description" : "Access token missing or invalid"
          },
          "404" : {
            "description" : "selected user not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "User" : {
        "type" : "object",
        "properties" : {
          "userName" : {
            "type" : "string",
            "example" : "Max Mustermann"
          },
          "experienceLevel" : {
            "type" : "integer",
            "example" : 0
          },
          "points" : {
            "type" : "integer",
            "example" : 0
          }
        },
        "description" : "Model for User",
        "example" : {
          "experienceLevel" : 0,
          "userName" : "Max Mustermann",
          "points" : 0
        },
        "xml" : {
          "name" : "user"
        }
      },
...
      "users_body" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "example" : "Max Mustermann"
          },
          "password" : {
            "type" : "string",
            "format" : "password",
            "example" : "Password"
          }
        }
      },
      "users_userId_body" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string",
            "format" : "password"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}